+++ MultTests.cpp
@@ -17,16 +17,16 @@ bool DoMultTest(int64_t a, int64_t b)
 
     bool passed = true;
 
-    if ((long long)csi != c)
+    if ((int64_t)csi != c)
     {
         std::cout << "* FAIL: " << asi.ToString() << " * " << bsi.ToString() << " = " << csi.ToString()
             << " should be " << c << std::endl;
         passed = false;
     }
 
-    int64_t d = (long long)dsi;
+    int64_t d = (int64_t)dsi;
 
-    if ((long long)dsi != c)
+    if ((int64_t)dsi != c)
     {
         std::cout << "*= FAIL: " << asi.ToString() << " *= " << bsi.ToString() << " makes " << dsi.ToString()
             << "should be " << c << std::endl;
